refactor(make): add support for common/*.S files
